+++ actix-server/CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add `ServerBuilder::graceful_shutdown_signal()` to propagate graceful shutdown to services.
 - Minimum supported Rust version (MSRV) is now 1.88.
 - internal: Use `core::future::{ready, Ready}` instead of actix-utils'
 - actix-server no longer removes the uds socket [#442]